+++ src/op/is_semicomplete.rs
@@ -21,19 +21,62 @@
 //! assert!(AdjacencyList::random_tournament(3, 0).is_semicomplete());
 //! ```
 
-use crate::{
-    HasArc,
-    Order,
-};
-
 /// Check whether a digraph is semicomplete.
 ///
 /// # Implementing [`IsSemicomplete`] for a custom type
 ///
 /// Provide an implementation of
 /// [`is_semicomplete`](IsSemicomplete::is_semicomplete) that returns whether
-/// the digraph is semicomplete OR implement [`HasArc`] and [`Order`].
+/// the digraph is semicomplete.
 ///
+/// ```
+/// use {
+///     graaf::IsSemicomplete,
+///     std::collections::BTreeSet,
+/// };
+///
+/// struct AdjacencyList {
+///     pub arcs: Vec<BTreeSet<usize>>,
+/// }
+///
+/// impl IsSemicomplete for AdjacencyList {
+///     fn is_semicomplete(&self) -> bool {
+///         let order = self.arcs.len();
+///
+///         (0..order).all(|u| {
+///             (u + 1..order).all(|v| {
+///                 self.arcs[u].contains(&v) || self.arcs[v].contains(&u)
+///             })
+///         })
+///     }
+/// }
+///
+/// assert!(AdjacencyList {
+///     arcs: vec![
+///         BTreeSet::from([1, 2]),
+///         BTreeSet::from([0, 2]),
+///         BTreeSet::from([0, 1]),
+///     ]
+/// }
+/// .is_semicomplete());
+///
+/// assert!(AdjacencyList {
+///     arcs: vec![
+///         BTreeSet::from([1, 2]),
+///         BTreeSet::from([0, 2]),
+///         BTreeSet::from([0]),
+///     ]
+/// }
+/// .is_semicomplete());
+///
+/// assert!(!AdjacencyList {
+///     arcs: vec![BTreeSet::new(); 3]
+/// }
+/// .is_semicomplete());
+/// ```
+///
+/// Implementations can be built with the [`HasArc`](crate::HasArc) and
+/// [`Order`](crate::Order) traits.
 /// ```
 /// use {
 ///     graaf::{
@@ -63,6 +106,17 @@ use crate::{
 ///     }
 /// }
 ///
+/// impl IsSemicomplete for AdjacencyList {
+///     fn is_semicomplete(&self) -> bool {
+///         let order = self.order();
+///
+///         (0..order).all(|u| {
+///             (u + 1..order)
+///                 .all(|v| self.has_arc(u, v) || self.has_arc(v, u))
+///         })
+///     }
+/// }
+///
 /// assert!(AdjacencyList {
 ///     arcs: vec![
 ///         BTreeSet::from([1, 2]),
@@ -86,8 +140,6 @@ use crate::{
 /// }
 /// .is_semicomplete());
 /// ```
-///
-/// # Examples
 pub trait IsSemicomplete {
     /// Check whether the digraph is semicomplete.
     ///
@@ -108,17 +160,4 @@ pub trait IsSemicomplete {
     /// ```
     #[must_use]
     fn is_semicomplete(&self) -> bool;
-}
-
-impl<D> IsSemicomplete for D
-where
-    D: HasArc + Order,
-{
-    fn is_semicomplete(&self) -> bool {
-        let order = self.order();
-
-        (0..order).all(|u| {
-            (u + 1..order).all(|v| self.has_arc(u, v) || self.has_arc(v, u))
-        })
-    }
 }