+++ src/repr/adjacency_list_weighted/mod.rs
@@ -41,6 +41,7 @@ use {
         HasEdge,
         Indegree,
         IsComplete,
+        IsSemicomplete,
         IsSimple,
         IsTournament,
         Order,
@@ -239,6 +240,18 @@ impl<W> IsComplete for AdjacencyListWeighted<W> {
 
         self.size() == order * (order - 1)
             && (0..order).all(|u| (u + 1..order).all(|v| self.has_edge(u, v)))
+    }
+}
+
+impl<W> IsSemicomplete for AdjacencyListWeighted<W> {
+    fn is_semicomplete(&self) -> bool {
+        let order = self.order();
+
+        self.size() >= order * (order - 1) / 2
+            && (0..order).all(|u| {
+                (u + 1..order)
+                    .all(|v| self.has_arc(u, v) || self.has_arc(v, u))
+            })
     }
 }