+++ src/repr/adjacency_matrix/mod.rs
@@ -130,6 +130,7 @@ use crate::{
     HasArc,
     Indegree,
     IsComplete,
+    IsSemicomplete,
     IsSimple,
     IsTournament,
     Order,
@@ -646,6 +647,18 @@ impl Indegree for AdjacencyMatrix {
 impl IsComplete for AdjacencyMatrix {
     fn is_complete(&self) -> bool {
         *self == Self::complete(self.order())
+    }
+}
+
+impl IsSemicomplete for AdjacencyMatrix {
+    fn is_semicomplete(&self) -> bool {
+        let order = self.order();
+
+        self.size() >= order * (order - 1) / 2
+            && (0..order).all(|u| {
+                (u + 1..order)
+                    .all(|v| self.has_arc(u, v) || self.has_arc(v, u))
+            })
     }
 }