+++ src/repr/edge_list/mod.rs
@@ -122,6 +122,7 @@ use {
         HasArc,
         Indegree,
         IsComplete,
+        IsSemicomplete,
         IsSimple,
         IsTournament,
         Order,
@@ -550,6 +551,18 @@ impl Indegree for EdgeList {
 impl IsComplete for EdgeList {
     fn is_complete(&self) -> bool {
         *self == Self::complete(self.order())
+    }
+}
+
+impl IsSemicomplete for EdgeList {
+    fn is_semicomplete(&self) -> bool {
+        let order = self.order();
+
+        self.size() >= order * (order - 1) / 2
+            && (0..order).all(|u| {
+                (u + 1..order)
+                    .all(|v| self.has_arc(u, v) || self.has_arc(v, u))
+            })
     }
 }